refactor: move theme file
