+++ src/pages/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import { ThemeProvider } from '@material-ui/core/styles'
 import { Helmet } from 'react-helmet'
-import { darkTheme, lightTheme } from './theme'
+import { darkTheme, lightTheme } from '../config/theme'
 import Header from '../components/Header'
 import icon from '../static/Icon.png'
 import Footer from '../components/Footer'
